refactor: dissolve governance/core_write.cpp into class home files
Unlike the evo and llmq JSON writers, nothing that links only
libbitcoin_common prints governance objects: dash-tx's TxToUniv() knows no
governance payload, and all callers of these ToJson() definitions sit in
libbitcoin_node. Keeping them in a libbitcoin_common file bought nothing and
forced governance/governance.h into the core_io module, creating the circular
dependency core_io -> governance/governance -> governance/superblock ->
core_io.

Move CGovernanceManager::ToJson() to governance/governance.cpp,
CGovernanceObject::GetInnerJson()/GetVotesJson() to governance/object.cpp and
Governance::Object::ToJson() to governance/common.cpp, and delete the file
along with its linter module mapping and the now-cleared suppression.
