+++ contrib/devtools/circular-dependencies.py
@@ -12,7 +12,6 @@ MAPPING = {
     'core_read.cpp': 'core_io.cpp',
     'core_write.cpp': 'core_io.cpp',
     'evo/core_write.cpp': 'core_io.cpp',
-    'governance/core_write.cpp': 'core_io.cpp',
     'llmq/core_write.cpp': 'core_io.cpp',
 }
